requirements.txt
python-dotenv
backports.functools-lru-cache==2.0.0
bcrypt==4.3.0
blinker==1.9.0
caniusepython3==7.3.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
distlib==0.3.9
et_xmlfile==2.0.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fpdf==1.7.2
greenlet==3.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
git+https://github.com/dmwangike/jinjasql.git@master
jwt==1.3.1
#MarkupSafe==3.0.2
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psycopg2-binary
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==8.2
reportlab==4.4.1
requests==2.32.3
setuptools==80.8.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1
